requirements.txt
fastapi
uvicorn[standard]
aiohttp
websockets
pandas
numpy
aioredis
gql
pydantic
